+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.scss';
-import { Clock } from './Clock';
+import { Clock } from './Components/Clock';
 
 interface State {
   isClockVisible: boolean,
